ADDS: Rooms!
Users now register via a name & the room they wish to join. Rooms are
concurrently handled by a worker pool which continuially checks the
existing rooms for broadcasts and sends workers off to handle the
messages before rejoining the pool.
Also lots of refactoring and a stress test which required the server to
run at the same time.